requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
cx_Oracle==8.3.0
